Adds support for switching browser page contexts.
The user can now specify 3 different statements for switching page
contexts (2 accomplish the same goal). This feature accounts for both
cases:
    - the change is explicit (now i want to navigate to this page)
    - implicit (a clicked link got me here).

Please refer to the docs for more information.

Excerpt from the docs:

---------

Allow the user to specify that the browser has implicitly navigated
to a new page (usually by clicking a link or submitting a form).

    When the browser is now at "Requisitions"
    # or
    Given the browser is now at "Requisitions"

This step simply changes the context of the browser page to allow the user
to specify elements using the page's convenience selectors.

In some cases it is not possible to assert that the page URL is some value
because the value is determined at runtime (dynamic URLs, etc). In most cases
users should prefer the assertion statement:

    Then the browser should be at "Requisitions"

This statement not only switches the context, but asserts that the current URL
is correct for the given page context.