Implements sentences for sending large text to elements
Specifying text to enter into an editor can make the feature files too
verbose and distract from the intented goal of the assertion. Therefore,
a sentece has been implmented to allow content such as markdown or JSON
to be loaded from the resources directory and entered into a text area.