Page Element Find is much more Robust Now.
A number of checks have been added to the find method on a page. Before
selecting an item, the page will check if the element exists, and wait up a few seconds for it to render if it isn't.

The page also does a lot more verification to ensure that the user is
getting the fetched results they expect. The Page now enforces a single
result be returned from a fetch unless otherwise specified, and failing
to return anything raises a custom NoElementFound exception for easy
debugging.

Globally, a new decorator is introduced that attempts to handle browser
based network failures or slow loads. The `allow_retries` decorator will
catch any of the provided exceptions (none by default) and if an
exception of that type is encountered, the function or method will be
retried up to a certain, configurable, amount of times before raising
the original exception.

This should hopefully help with testing in the real world.