Convert veripy to Python module
Due to the refactor, the project now has a very nested project structure
which introduces odd behavior when importing via a modified PYTHONPATH
environment setting. As such, the project has been converted into an
installable step library.

This now has the following benefit of client bundles no longer have
to be embedded into veripy in order to trick Behave into searching
veripy for required files. The only the drawback is the the client
bundle must now specify a steps directory and environment file, as is
required by Behave. Even though this is drawback, the benefit is if
clients require their own special senentences or environements, it is
now trivial to override Veripy/Behave functiontionality.