Fixes behavior of clicking on Nth element
This fix reverts clicking of Nth element to it's previous behavior of
returning multiple elements and then selecting the Nth option in that
list. The user must remember to enable the return of multiple matches in
the page fixutre in order for this step to work properly.

The reason for this fix is becuase in its current state, the xpath must
be specified in the step, which may be too difficult for
non-programmers.