Refactor sentences to be present tense
This change refactors all sentences to be present tense. For example,
instead of `Then the browser should be at "Home Page"`, it is now
written as "THen the browser is at "Home Page"`.

Additionally, all references to "the user" have been removed in order
for the sentences to read more like instrustions given to a person.

Finally, all parameters have been normalized to be more consistent and
descriptive (e.g. "name" to "page_name", "element" to "element_name")

Other minor changes include changes to the docs and adding a service to
readily view generated docs.