Fixes to how select options are found.
We've elected to not use the provided splinter `select()` method because it
makes the assumption that the select element will have a `name` attribute which is unfortunately not a sound assumption. In this case, we use the select to get all of the options that are children of the select w/o caring about how the select is constructed.

The only limitation of this method is that the select's SELECTOR in the
configuration file *MUST* be an XPath. There's no easy way to do this
for other selector types w/o lots of weird string manipluation.

- Adds a new public method to the Page to allow users to fetch the selector
properties by name.

- Adds a new exceptions file and `ImproperlyConfigured` exception for
configuration issues. (See Docs)